remove redundant type conv

+++ service/promotion/by_type.go
@@ -29,7 +29,7 @@ const (
 )
 
 func RewardByType(c context.Context, p models.Promotion, s models.PromotionSession, userID, progress int64, now time.Time) (reward, meetGapType int64, vipIncrementDetail models.VipIncrementDetail, err error) {
-	switch int64(p.Type) {
+	switch p.Type {
 	case models.PromotionTypeVipBirthdayB:
 		err := cache.RedisStore.Get(fmt.Sprintf(birthdayBonusRewardCacheKey, userID), &reward)
 		if errors.Is(err, persist.ErrCacheMiss) {
@@ -58,7 +58,7 @@ func RewardByType(c context.Context, p models.Promotion, s models.PromotionSessi
 
 func ProgressByType(ctx context.Context, p models.Promotion, s models.PromotionSession, userID int64, now time.Time) (progress int64) {
 	//ctx = contextify.AppendCtx(ctx, contextify.DefaultContextKey, fmt.Sprintf("ProgressByType() p.Type %d", p.Type))
-	switch int64(p.Type) {
+	switch p.Type {
 	// not necessary
 	// case models.PromotionTypeVipReferral, models.PromotionTypeVipRebate:
 	// 	//separate handling based on separate table
@@ -105,9 +105,9 @@ func ProgressByType(ctx context.Context, p models.Promotion, s models.PromotionS
 func ClaimStatusByType(c context.Context, p models.Promotion, s models.PromotionSession, userID int64, now time.Time) (claim serializer.ClaimStatus) {
 	claim.ClaimStart = s.ClaimStart.Unix()
 	claim.ClaimEnd = s.ClaimEnd.Unix()
-	switch int64(p.Type) {
+	switch p.Type {
 	case models.PromotionTypeVipRebate, models.PromotionTypeVipPromotionB, models.PromotionTypeVipWeeklyB, models.PromotionTypeVipBirthdayB:
-		v, err := model.VoucherGetByUniqueID(c, models.GenerateVoucherUniqueId(int64(p.Type), p.ID, s.ID, userID, 0, buildSuffixByType(c, p, userID)))
+		v, err := model.VoucherGetByUniqueID(c, models.GenerateVoucherUniqueId(p.Type, p.ID, s.ID, userID, 0, buildSuffixByType(c, p, userID)))
 		if err == nil && v.ID != 0 {
 			claim.HasClaimed = true
 		}
@@ -134,13 +134,13 @@ func ClaimVoucherByType(c context.Context, p models.Promotion, s models.Promotio
 	voucher = CraftVoucherByType(c, p, s, v, rewardAmount, userID, promotionRequestID, now, meetGapType, vipIncrementDetail)
 	lang := model.GetUserLang(userID)
 
-	switch int64(p.Type) {
+	switch p.Type {
 	case models.PromotionTypeFirstDepB, models.PromotionTypeReDepB, models.PromotionTypeBeginnerB, models.PromotionTypeOneTimeDepB:
 		//add money and insert voucher
 		// add cash order
 		err = model.DB.Clauses(dbresolver.Use("txConn")).Debug().WithContext(c).Transaction(func(tx *gorm.DB) error {
 			wagerChange := voucher.WagerMultiplier * rewardAmount
-			err = CreateCashOrder(tx, int64(p.Type), userID, rewardAmount, wagerChange, "", "")
+			err = CreateCashOrder(tx, p.Type, userID, rewardAmount, wagerChange, "", "")
 			if err != nil {
 				return err
 			}
@@ -148,13 +148,13 @@ func ClaimVoucherByType(c context.Context, p models.Promotion, s models.Promotio
 			if err != nil {
 				return err
 			}
-			if int64(p.Type) == models.PromotionTypeBeginnerB {
+			if p.Type == models.PromotionTypeBeginnerB {
 				err = model.CreateUserAchievement(userID, models.UserAchievementIdFirstAppLoginReward)
 				if err != nil {
 					return err
 				}
 			}
-			if int64(p.Type) == models.PromotionTypeOneTimeDepB {
+			if p.Type == models.PromotionTypeOneTimeDepB {
 				err = model.CreateUserAchievement(userID, models.UserAchievementIdFirstDepositBonusReward)
 				if err != nil {
 					return err
@@ -173,7 +173,7 @@ func ClaimVoucherByType(c context.Context, p models.Promotion, s models.Promotio
 	case models.PromotionTypeVipBirthdayB:
 		err = model.DB.Clauses(dbresolver.Use("txConn")).Debug().WithContext(c).Transaction(func(tx *gorm.DB) error {
 			wagerChange := voucher.WagerMultiplier * rewardAmount
-			err = CreateCashOrder(tx, int64(p.Type), userID, rewardAmount, wagerChange, "", "")
+			err = CreateCashOrder(tx, p.Type, userID, rewardAmount, wagerChange, "", "")
 			if err != nil {
 				return err
 			}
@@ -189,7 +189,7 @@ func ClaimVoucherByType(c context.Context, p models.Promotion, s models.Promotio
 	case models.PromotionTypeVipRebate, models.PromotionTypeVipPromotionB, models.PromotionTypeVipWeeklyB:
 		err = model.DB.Clauses(dbresolver.Use("txConn")).Debug().WithContext(c).Transaction(func(tx *gorm.DB) error {
 			wagerChange := voucher.WagerMultiplier * rewardAmount
-			err = CreateCashOrder(tx, int64(p.Type), userID, rewardAmount, wagerChange, "", "")
+			err = CreateCashOrder(tx, p.Type, userID, rewardAmount, wagerChange, "", "")
 			if err != nil {
 				return err
 			}
@@ -206,7 +206,7 @@ func ClaimVoucherByType(c context.Context, p models.Promotion, s models.Promotio
 		})
 		if err == nil {
 			var nType, title, text string
-			switch int64(p.Type) {
+			switch p.Type {
 			case models.PromotionTypeVipRebate:
 				nType = consts.Notification_Type_Rebate
 				title = conf.GetI18N(lang).T(common.NOTIFICATION_REBATE_TITLE)
@@ -228,7 +228,7 @@ func ClaimVoucherByType(c context.Context, p models.Promotion, s models.Promotio
 	case models.PromotionTypeCustomTemplate:
 		err = model.DB.Clauses(dbresolver.Use("txConn")).Debug().WithContext(c).Transaction(func(tx *gorm.DB) error {
 			wagerChange := voucher.WagerMultiplier * rewardAmount
-			err = CreateCashOrder(tx, int64(p.Type), userID, rewardAmount, wagerChange, "", p.Name)
+			err = CreateCashOrder(tx, p.Type, userID, rewardAmount, wagerChange, "", p.Name)
 			if err != nil {
 				return err
 			}
@@ -243,7 +243,7 @@ func ClaimVoucherByType(c context.Context, p models.Promotion, s models.Promotio
 }
 
 func ExtraByType(c context.Context, p models.Promotion, s models.PromotionSession, userID, progress int64, now time.Time) (extra any) {
-	switch int64(p.Type) {
+	switch p.Type {
 	case models.PromotionTypeVipReferral:
 		// Get next session's reward - actually just records that cannot be claimed yet
 		summaries, err := model.GetReferralAllianceSummaries(model.GetReferralAllianceSummaryCond{
@@ -312,7 +312,7 @@ func CraftVoucherByType(c context.Context, p models.Promotion, s models.Promotio
 	status := models.VoucherStatusReady
 	isUsable := false
 	suffix := buildSuffixByType(c, p, userID)
-	switch int64(p.Type) {
+	switch p.Type {
 	case models.PromotionTypeFirstDepB, models.PromotionTypeReDepB, models.PromotionTypeBeginnerB, models.PromotionTypeVipReferral, models.PromotionTypeVipRebate, models.PromotionTypeVipPromotionB, models.PromotionTypeVipWeeklyB, models.PromotionTypeVipBirthdayB, models.PromotionTypeCustomTemplate:
 		status = models.VoucherStatusRedeemed
 	case models.PromotionTypeFirstDepIns, models.PromotionTypeReDepIns:
@@ -321,7 +321,7 @@ func CraftVoucherByType(c context.Context, p models.Promotion, s models.Promotio
 
 	voucherName := model.AmountReplace(v.Name, float64(rewardAmount)/100)
 	wagerMultiplier := v.WagerMultiplier
-	switch int64(p.Type) {
+	switch p.Type {
 	case models.PromotionTypeVipRebate:
 		vip, _ := model.GetVipWithDefault(c, userID)
 		wagerMultiplier = vip.VipRule.BirthdayBenefitWagerMultiplier
@@ -462,7 +462,7 @@ func claimVoucherReferralVip(c context.Context, p models.Promotion, voucher mode
 		})
 
 		wagerChange := user.ReferralWagerMultiplier * totalClaimable
-		err = CreateCashOrder(tx, int64(p.Type), user.ID, totalClaimable, wagerChange, cashOrderNotes, "")
+		err = CreateCashOrder(tx, p.Type, user.ID, totalClaimable, wagerChange, cashOrderNotes, "")
 		if err != nil {
 			return fmt.Errorf("failed to create cash order: %w", err)
 		}
@@ -536,7 +536,7 @@ func buildSuffixByType(c context.Context, p models.Promotion, userID int64) stri
 	today := Today0am()
 	suffix := ""
 	vip, _ := model.GetVipWithDefault(c, userID)
-	switch int64(p.Type) {
+	switch p.Type {
 	case models.PromotionTypeVipRebate:
 		suffix = fmt.Sprintf("date-%s", today.Format(time.DateOnly))
 	case models.PromotionTypeVipWeeklyB:
